+++ FlowNet2_src/models/components/ops/correlation/make.sh
@@ -5,9 +5,8 @@ TORCH=$($PYTHON -c "import os; import torch; print(os.path.dirname(torch.__file_
 cd src
 
 echo "Compiling correlation kernels by nvcc..."
-
-rm correlation_cuda_kernel.o
-rm -r ../_ext
+[ -f correlation_cuda_kernel.o ] && rm correlation_cuda_kernel.o
+[ -d ../_ext ] && rm -r ../_ext
 
 nvcc -c -o correlation_cuda_kernel.o correlation_cuda_kernel.cu -x cu -Xcompiler -fPIC -arch=sm_30
